Travis-CI CMake workaround
Since we are integrating dependencies using add_subdirectory().
We should add them to install(TARGETS) but we can't.

Until cmake 3.13 it is impossible to add target to install which has not been declare in the current scope i.e. CMakeLists.txt
It is also impossible to install target alias so you must retrieve the ALIASED_TARGET if any...