Add special error message for .cpp SOURCE
Before this change, when a user of an or-tools release tries to use a
".cpp" file with e.g.:

    make run SOURCE=./foo.cpp

The error message is:

    make: *** No rule to make target 'run'.  Stop.

This message does not really help to understand that the makefile is
designed to recognize only ".cc" files as C++ sources.

This change adds a specific error message when the source is a ".cpp"
file, inviting the user to use ".cc" instead.