Use bazel workspace target for zlib
Fixes https://github.com/google/or-tools/issues/3060

Without this, linux builds will fail at runtime if z.so is not
installed on the deploy platform. This is problematic for example, when
using rules_docker to create a minimal image for a cc_binary to run.

NOTE: Users who want to use a system-installed version of zlib can still
do so by creating a library in their workspace called `@zlib:zlib` with
`-lz` in its linkopts. This gives flexibility at the workspace level for
choosing how to build zlib.